Moved Point3D to upper level with multiple changes
- Point3D is now immutable (removed setX(), setY(), and setZ())
- x, y, and z is now protected instead of private
- separated distance() into both distance() and distanceSquared()
- deprecated no-arg constructor and made static Point3D.ZERO variable.
- made a 2D constructor (initializing only the X and Y values)
- added documentation to every function
